requirements.txt
fastapi~=0.115.13

openai~=1.91.0
anthropic

python-dotenv~=1.1.1
pydantic~=2.11.7

requests~=2.32.3